requirements.txt
discord.py>=2.3
SQLAlchemy>=2.0
PyYAML>=6.0
python-dotenv>=1.0
psycopg2-binary>=2.9
pytest>=8.0
pytest-asyncio>=0.23
